+++ alita/modules/purges.go
@@ -158,6 +158,19 @@ func (m moduleStruct) purge(bot *gotgbot.Bot, ctx *ext.Context) error {
 		msgId := msg.ReplyToMessage.MessageId
 		deleteTo := msg.MessageId - 1
 		totalMsgs := deleteTo - msgId + 1 // adding 1 because we want to delete the message we are replying to
+
+		// Limit purge range to prevent abuse and API overload
+		const maxPurgeMessages = 1000
+		if totalMsgs > maxPurgeMessages {
+			tr := i18n.MustNewTranslator(db.GetLanguage(ctx))
+			text, _ := tr.GetString("purges_limit_exceeded")
+			_, err := msg.Reply(bot, fmt.Sprintf(text, maxPurgeMessages), helpers.Shtml())
+			if err != nil {
+				log.Error(err)
+			}
+			return ext.EndGroups
+		}
+
 		purge := m.purgeMsgs(bot, chat, false, msgId, deleteTo)
 		_ = helpers.DeleteMessageWithErrorHandling(bot, chat.Id, msg.MessageId)
